rename tightly coupled component files
