+++ src/lib/games/quantum-tictactoe/GameBoardSquareMarkClassical.svelte
